Add authentication middleware to feed-parser, refator and udpate test coverage to 100%
