+++ src/api/feed-discovery/jest.config.js
@@ -5,4 +5,5 @@ module.exports = {
   rootDir: '../../..',
   testMatch: ['<rootDir>/src/api/feed-discovery/test/*.test.js'],
   collectCoverageFrom: ['<rootDir>/src/api/feed-discovery/src/**/*.js'],
+  testEnvironment: 'node',
 };
